ShoppingComplex---E-Commerce-Website-main/ShoppingComplex---E-Commerce-Website-main/Backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1
python-dotenv==1.0.0
pymongo==4.5.0
sqlalchemy==2.0.21
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
requests==2.31.0
